Fix introspected benchmark
