+++ benchmarks/src/jmh/java/io/micronaut/deserialization/xml/IntrospectedBook.java
@@ -1,0 +1,117 @@
+/*
+ *
+ *  * Copyright 2017-2019 original authors
+ *  *
+ *  * Licensed under the Apache License, Version 2.0 (the "License");
+ *  * you may not use this file except in compliance with the License.
+ *  * You may obtain a copy of the License at
+ *  *
+ *  * http://www.apache.org/licenses/LICENSE-2.0
+ *  *
+ *  * Unless required by applicable law or agreed to in writing, software
+ *  * distributed under the License is distributed on an "AS IS" BASIS,
+ *  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  * See the License for the specific language governing permissions and
+ *  * limitations under the License.
+ *
+ */
+
+package io.micronaut.deserialization.xml;
+
+import io.micronaut.context.annotation.Prototype;
+import io.micronaut.core.annotation.Introspected;
+
+@Introspected
+@Prototype
+public class IntrospectedBook {
+    IntrospectedAuthor author;
+    String bookName;
+    String field1;
+    String field2;
+    String field3;
+    String field4;
+    String field5;
+    String field6;
+    String field7;
+    String field8;
+
+    public IntrospectedAuthor getAuthor() {
+        return this.author;
+    }
+
+    public void setAuthor(IntrospectedAuthor author) {
+        this.author = author;
+    }
+
+    public String getBookName() {
+        return this.bookName;
+    }
+
+    public void setBookName(String bookName) {
+        this.bookName = bookName;
+    }
+
+    public String getField1() {
+        return this.field1;
+    }
+
+    public void setField1(String field1) {
+        this.field1 = field1;
+    }
+
+    public String getField2() {
+        return this.field2;
+    }
+
+    public void setField2(String field2) {
+        this.field2 = field2;
+    }
+
+    public String getField3() {
+        return this.field3;
+    }
+
+    public void setField3(String field3) {
+        this.field3 = field3;
+    }
+
+    public String getField4() {
+        return this.field4;
+    }
+
+    public void setField4(String field4) {
+        this.field4 = field4;
+    }
+
+    public String getField5() {
+        return this.field5;
+    }
+
+    public void setField5(String field5) {
+        this.field5 = field5;
+    }
+
+    public String getField6() {
+        return this.field6;
+    }
+
+    public void setField6(String field6) {
+        this.field6 = field6;
+    }
+
+    public String getField7() {
+        return this.field7;
+    }
+
+    public void setField7(String field7) {
+        this.field7 = field7;
+    }
+
+    public String getField8() {
+        return this.field8;
+    }
+
+    public void setField8(String field8) {
+        this.field8 = field8;
+    }
+}
